a2a-weather-demo/requirements.txt
gooogle-adk==<latest>
litellm==<latest>
python-dotenv
rich
prompt_toolkit